refactor: moved custom navigation into hook
